pyproject.toml
[project]
name = "viking"
version = "0.1.0"
description = "AI-powered menu picker for panel.kuchniavikinga.pl"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.7",
    "typer>=0.12",
    "python-dotenv>=1.0",
    "openai>=1.40",
    "anthropic>=0.34",
    "rich>=13.7",
]

[project.optional-dependencies]
playwright = ["playwright>=1.44"]
dev = ["pytest>=8.0", "pytest-mock>=3.12"]

[project.scripts]
viking = "viking.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/viking"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
